refactored task handler
